Pages moved to mustache way Defualt timeout increased to 90seconds
