--- packages/perfrunner-reporters/src/reporters/html/renders/index.ts
@@ -1,4 +1,0 @@
-import { renderPages } from './page';
-import { renderPlacholdersForCharts } from './placeholders';
-
-export { renderPlacholdersForCharts, renderPages };
